attestation: register extension IDs/masks; widen masks to uint64_t (M2-1)
Add EXT_ID constants (29-33) and update ALLOWED_EXTS masks for the five
attestation extensions per draft-fossati-seat-early-attestation-03:

  OF_CH/OF_EE: EVIDENCE_REQUEST, EVIDENCE_PROPOSAL,
               RESULTS_REQUEST, RESULTS_PROPOSAL
  OF_CT:       ATTESTATION

Also wire all five into mbedtls_ssl_get_extension_id(), the name table,
and the type table for debug logging.

The addition of IDs 29-33 exhausted the uint32_t range of the existing
extension mask infrastructure (EXT_MASK uses 1ULL but fields were
uint32_t). Widen consistently:

  - ssl_misc.h: sent_extensions, received_extensions fields
  - ssl_misc.h: get_extension_mask() and check_received_extension()
    declarations
  - ssl_tls.c:  get_extension_mask() definition; print_extensions()
  - ssl_debug_helpers.h: print_extensions() declaration
  - ssl_tls13_generic.c: check_received_extension() definition
  - ssl_tls13_server.c: allowed_exts local; exts_mask param/masked local

SSL test suite passes (ctest -R ssl).

+++ library/ssl_debug_helpers.h
@@ -38,7 +38,7 @@ const char *mbedtls_ssl_get_extension_name(unsigned int extension_type);
 
 void mbedtls_ssl_print_extensions(const mbedtls_ssl_context *ssl,
                                   int level, const char *file, int line,
-                                  int hs_msg_type, uint32_t extensions_mask,
+                                  int hs_msg_type, uint64_t extensions_mask,
                                   const char *extra);
 
 void mbedtls_ssl_print_extension(const mbedtls_ssl_context *ssl,
